feat: react-native-worklets from bundle

+++ packages/react-native-worklets/plugin/src/workletFactory.ts
@@ -44,7 +44,7 @@ import {
   variableDeclarator,
 } from '@babel/types';
 import { strict as assert } from 'assert';
-import { mkdirSync, writeFileSync } from 'fs';
+import { mkdirSync, writeFileSync, existsSync } from 'fs';
 import { basename, dirname, relative, resolve } from 'path';
 
 import { globals } from './globals';
@@ -355,10 +355,7 @@ export function makeWorkletFactory(
       importDeclaration(
         [cloneNode(binding.path.node as ImportSpecifier, true)],
         stringLiteral(
-          require.resolve(
-            (binding.path.parentPath!.node as ImportDeclaration).source.value,
-            { paths: [dirname(state.file.opts.filename!)] }
-          )
+          (binding.path.parentPath!.node as ImportDeclaration).source.value
         )
       )
     );
@@ -389,16 +386,33 @@ export function makeWorkletFactory(
   );
 
   try {
-    mkdirSync(filesDirPath, {});
+    if (!existsSync(filesDirPath)) {
+      mkdirSync(filesDirPath, {});
+    }
   } catch (e) {
     // Nothing.
   }
 
   const dedicatedFilePath = resolve(filesDirPath, `${workletHash}.js`);
 
+  // @ts-expect-error wwww
+  if (!state.file.metadata.virtualModules) {
+    // @ts-expect-error wwww
+    state.file.metadata.virtualModules = new Map();
+  }
+  // @ts-expect-error wwww
+  state.file.metadata.virtualModules.set(dedicatedFilePath, newProg);
+
   try {
-    writeFileSync(dedicatedFilePath, transformedProg);
+    // If a file exists, abort the process.
+    if (!existsSync(dedicatedFilePath)) {
+      // temporary
+      writeFileSync(dedicatedFilePath, transformedProg);
+      console.error('Saved worklet to file ', dedicatedFilePath);
+      // console.error('babel time', new Date().toISOString());
+    }
   } catch (_e) {
+    console.error('Error while writing worklet to file: ', _e);
     // Nothing.
   }
